+++ src/com/maddyhome/idea/vim/extension/commentary/CommentaryExtension.java
@@ -48,6 +48,11 @@ public class CommentaryExtension extends VimNonDisposableExtension {
 
   private static class CommentMotionHandler implements VimExtensionHandler {
     @Override
+    public boolean isRepeatable() {
+      return true;
+    }
+
+    @Override
     public void execute(@NotNull Editor editor, @NotNull DataContext context) {
       setOperatorFunction(new Operator());
       executeNormal(parseKeys("g@"), editor);
@@ -128,6 +133,11 @@ public class CommentaryExtension extends VimNonDisposableExtension {
   }
 
   private static class CommentLineHandler implements VimExtensionHandler {
+    @Override
+    public boolean isRepeatable() {
+      return true;
+    }
+
     @Override
     public void execute(@NotNull Editor editor, @NotNull DataContext context) {
       final int offset = editor.getCaretModel().getOffset();